Complete function getAllCargo in controller,  ceate page show-cargos, page home-manager, modify page home, create function deleteCargo in controller

+++ src/main/java/org/apptopia/waterwayfreightsystem/api/api/application/usecases/cargo/RawCargoInput.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Data @AllArgsConstructor
 @NoArgsConstructor
 public class RawCargoInput implements UseCaseInput {
-	private Integer idCargo;
+	private Long idCargo;
 	private Long idOwner;
 	private Float cargoWeight;
 }
